Add README.md for project setup
A README.md file has been created to aid with the setup of the Next.js project. This includes instructions on how to run the development server and guidance on how to deploy the Next.js app on Vercel. Alongside this, resources for learning more about Next.js have been provided.